% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{eggan_ghl}
\alias{eggan_ghl}
\title{EggaN Greenland halibut survey data}
\format{
A dataframe
}
\source{
Institute of Marine Research (\url{https://www.imr.no})
}
\usage{
data(eggan_ghl)
}
\description{
EggaN Greenland halibut survey data
}
\keyword{datasets}
