% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{eggan_ghl_original_strata}
\alias{eggan_ghl_original_strata}
\title{EggaN original strata areas for Greenland halibut}
\format{
A dataframe
}
\source{
Institute of Marine Research (\url{https://www.imr.no})
}
\usage{
data(eggan_ghl_original_strata)
}
\description{
EggaN original strata areas for Greenland halibut
}
\keyword{datasets}
