% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.strataPolygon.R
\name{plot.strataPolygon}
\alias{plot.strataPolygon}
\title{Plot a strata polygon object on a map}
\usage{
\method{plot}{strataPolygon}(
  x,
  basemap = TRUE,
  facetted = FALSE,
  fill_color = "red",
  fill_alpha = 0.8,
  ...
)
}
\arguments{
\item{x}{\code{strataPolygon} object from the \code{\link{strataPolygon}} function.}

\item{basemap}{Logical. Should the polygons be plotted on a \code{\link[ggOceanMaps]{basemap}} from the ggOceanMaps package.}

\item{facetted}{Logical indicating whether a facetted map should be returned instead of colored polygons overlaid on the same map.}

\item{...}{Additional parameters passed to \code{\link[ggOceanMaps]{basemap}}.}
}
\description{
Plot method for \code{\link[=strataPolygon]{strataPolygon}} objects.
}
\seealso{
\code{\link{strataPolygon}} \code{\link[ggOceanMaps]{basemap}}
}
\author{
Mikko Vihtakari
}
