% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStrata.R
\name{plotGeostrata}
\alias{plotGeostrata}
\title{Plot a geostrata data frame on a map}
\usage{
plotGeostrata(geostrata, plot = TRUE, ...)
}
\arguments{
\item{geostrata}{A data frame. See the same argument in \code{\link[=strataPolygon]{strataPolygon}}}

\item{plot}{Logical indicating whether ggOceanMaps plot should be returned. If \code{FALSE}, the geostrata polygons are returned as an \link[sf]{sf} objected instead.}

\item{...}{Additional parameters passed to \code{\link[ggOceanMaps]{basemap}}.}
}
\value{
If \code{plot = TRUE} a \link[ggplot2]{ggplot} map, which can be assigned to an object and modified as any ggplot object. Otherwise an \link[sf]{sfc}.
}
\description{
Plots geostrata argument from the \code{\link[=strataPolygon]{strataPolygon}} on a \code{\link[ggOceanMaps]{basemap}} together with bathymetry
}
\details{
This function is useful for fitting geostrata visually on a map before running the time-consuming \code{\link[=strataPolygon]{strataPolygon}} step. Can also be used to acquire \code{geostrata} as \link[sf]{sf} polygons
}
\author{
Mikko Vihtakari (Institute of Marine Research)
}
